% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{CV}
\alias{CV}
\title{Coefficient of variation}
\usage{
CV(mean, sd)
}
\arguments{
\item{x}{vector}
}
\description{
Compute coefficient ((sd/mean)*100) of variation of a vector;
}
\keyword{coefficient}
\keyword{of}
\keyword{variation}

