% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{first.occ}
\alias{first.occ}
\title{First occurrence}
\usage{
first.occ(x)
}
\arguments{
\item{x}{vector}
}
\description{
Index of the first occurrence of unique values from a vector
}
\keyword{occurrence}

