% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{\%out\%}
\alias{\%out\%}
\title{%out%}
\usage{
x \%out\% y
}
\arguments{
\item{x}{Value or vector of interest}

\item{y}{Vector of values to compare to x}
}
\description{
Reverse of %in% - determine whether values are not in a vector
}
\examples{
x <- 11
y <- seq(1:10)
x\%out\%y
[1] TRUE
}
\keyword{out}

