% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{invlogit}
\alias{invlogit}
\title{Inverse logit}
\usage{
invlogit(x)
}
\arguments{
\item{x}{vector of logit transformed numerical values}
}
\description{
Function to compute inverse logit
}
\keyword{logistic}
\keyword{logit;}
\keyword{regression}

