% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{mean}
\alias{mean}
\title{Mean}
\usage{
mean(x)
}
\arguments{
\item{x}{Numeric vector}
}
\description{
Same as base mean, but defaults to printing 3 digits
}
\examples{
x <- rnorm(100)
mean(x)
}
\keyword{mean}

