% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{read.csvl}
\alias{read.csvl}
\title{Read csv lowercase}
\usage{
read.csvl(file)
}
\arguments{
\item{file}{filename of .csv file}
}
\description{
Modification of read.csv to automatically convert column names to all lowercase
}
\keyword{csv}
\keyword{read;}

