% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/theme_dark.R
\name{theme_dark}
\alias{theme_dark}
\title{theme_dark}
\usage{
theme_dark(base_size = 12, base_family = "")
}
\arguments{
\item{ggplot}{object}
}
\description{
Modification of theme_clark to add dark background & grey axis lines/text (based on solarized)
}
\keyword{theme}

