% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/theme_light.R
\name{theme_light}
\alias{theme_light}
\title{theme_light}
\usage{
theme_light(base_size = 12, base_family = "")
}
\arguments{
\item{ggplot}{object}
}
\description{
Modification of theme_clark to add dark background & grey axis lines/text (based on solarized)
}
\keyword{theme}

