% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{to_dev}
\alias{to_dev}
\title{General function to save figure files to different devices}
\usage{
to_dev(expr, dev, filename, ..., verbose = TRUE)
}
\arguments{
\item{expr}{figure file or function to create figure}

\item{dev}{device}

\item{filename}{filename for saved figure}

\item{...}{additional arguements passed to dev}
}
\description{
Adapted from nicercode.github.io
}
\keyword{figures}

