% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{to_png}
\alias{to_png}
\title{Function to save figure files as png}
\usage{
to_png(expr, filename, ...)
}
\arguments{
\item{expr}{figure file or function to create figure}

\item{filename}{filename for saved figure}

\item{...}{additional arguements passed to dev}
}
\description{
Adapted from nicercode.github.io
}
\keyword{figures}

